Check for file before deleting it
also commenting out pscp command

+++ Post-PublicIP.ps1
@@ -3,12 +3,17 @@ $webclient = New-Object System.Net.WebClient
 $ip = ($webclient.DownloadString($url)).Split(" ")[5].replace("</body>","").replace("</html>","")
 #print it
 $ip
+If( (Test-Path $env:USERPROFILE\MyIP.txt) -eq $true ) {
+    Remove-Item $env:USERPROFILE\MyIP.txt
+}
 
-Remove-Item $env:USERPROFILE\MyIP.txt
-Remove-Variable ip, webclient, url
 
 #write IP to file
 $ip | Out-File $env:USERPROFILE\MyIP.txt
 
+
+
 #send it via pscp
-C:\pscp.exe -pw <PASSWORD> -scp $env:USERPROFILE\MyIP.txt <USER>@<HOST>: 
+#C:\pscp.exe -pw <PASSWORD> -scp $env:USERPROFILE\MyIP.txt <USER>@<HOST>: 
+
+Remove-Variable ip, webclient, url
